Fix web panel settings, add range playback, harden gateway
Firmware:
- Accept numeric strings from the web settings form; ArduinoJson's operator|
  ignored them, so brightness, low-battery save, seek step, and sort changes
  were silently dropped.
- Honour HTTP byte ranges when serving recordings so Safari/iOS can play them
  and browsers can seek; answer 416 for unsatisfiable ranges.
- Refresh the panel UI: status badges, modified dates, storage/battery bars,
  file count with truncation notice, manual refresh, compact mobile layout.
- Cap voice-job poll read timeout at 10 s since it runs on the main loop.
- Use 64-bit arithmetic in WAV chunk bounds checks.

Gateway:
- Reject unknown voice profiles with 400 instead of silently using default.
- Persist the Codex turn id before honouring cancellation.
- Split long WAVs for whisper-server in a worker thread.
- Prune finished Codex jobs so long-running gateways do not grow unbounded.
- Add regression tests for profile validation and job pruning.

+++ firmware/src/media/wav_file.cpp
@@ -401,7 +401,8 @@ bool WavReader::begin(File file)
         }
         const std::uint32_t chunkSize = read32(chunk + 4);
         const std::uint32_t payloadOffset = offset + 8;
-        if (payloadOffset + chunkSize > file.size()) {
+        if (static_cast<std::uint64_t>(payloadOffset) + chunkSize >
+            file.size()) {
             error_ = WavError::kTruncated;
             return false;
         }
